requirements.txt
Flask==3.0.0
psycopg[binary]==3.2.4
python-dotenv==1.0.0
